renamme component
